chore: refactor imports and move constants
