+++ contracts/cosmwasm-vm/cw-ibc-core/src/constants.rs
@@ -1,0 +1,24 @@
+// Constants for Reply messages
+
+pub const EXECUTE_CREATE_CLIENT: u64 = 21;
+pub const EXECUTE_UPDATE_CLIENT: u64 = 22;
+pub const EXECUTE_UPGRADE_CLIENT: u64 = 23;
+pub const MISBEHAVIOUR: u64 = 23;
+
+pub const EXECUTE_CONNECTION_OPENTRY: u64 = 31;
+pub const EXECUTE_CONNECTION_OPENACK: u64 = 32;
+pub const EXECUTE_CONNECTION_OPENCONFIRM: u64 = 33;
+
+pub const EXECUTE_ON_CHANNEL_OPEN_INIT: u64 = 41;
+pub const EXECUTE_ON_CHANNEL_OPEN_TRY: u64 = 42;
+pub const EXECUTE_ON_CHANNEL_OPEN_TRY_ON_LIGHT_CLIENT: u64 = 421;
+pub const EXECUTE_ON_CHANNEL_OPEN_ACK_ON_LIGHT_CLIENT: u64 = 431;
+pub const EXECUTE_ON_CHANNEL_OPEN_ACK_ON_MODULE: u64 = 432;
+pub const EXECUTE_ON_CHANNEL_OPEN_CONFIRM_ON_LIGHT_CLIENT: u64 = 441;
+pub const EXECUTE_ON_CHANNEL_OPEN_CONFIRM_ON_MODULE: u64 = 442;
+pub const EXECUTE_ON_CHANNEL_CLOSE_INIT: u64 = 45;
+pub const EXECUTE_ON_CHANNEL_CLOSE_CONFIRM_ON_LIGHT_CLIENT: u64 = 461;
+pub const EXECUTE_ON_CHANNEL_CLOSE_CONFIRM_ON_MODULE: u64 = 462;
+
+pub const VALIDATE_ON_PACKET_TIMEOUT_ON_LIGHT_CLIENT: u64 = 54;
+pub const VALIDATE_ON_PACKET_TIMEOUT_ON_MODULE: u64 = 54;
