+++ contracts/cosmwasm-vm/cw-ibc-core/src/ics03_connection/mod.rs
@@ -18,6 +18,7 @@ use cosmwasm_std::Response;
 use cosmwasm_std::Storage;
 use cosmwasm_std::{from_binary, to_binary, to_vec, CosmosMsg, MessageInfo, Reply, SubMsg};
 
+pub use super::*;
 use crate::types::OpenConfirmResponse;
 use ibc::core::ics03_connection::connection::ConnectionEnd;
 use ibc::core::ics03_connection::error::ConnectionError;
