+++ contracts/cosmwasm-vm/cw-ibc-core/tests/channel/mod.rs
@@ -4,7 +4,8 @@ use crate::channel::test_timeout::get_dummy_raw_msg_timeout_on_close;
 use cosmwasm_std::from_binary;
 use cosmwasm_std::testing::mock_env;
 use cosmwasm_std::{Binary, IbcEndpoint, IbcPacket, IbcTimeout, IbcTimeoutBlock};
-use cw_ibc_core::ics04_channel::{compute_packet_commitment, PacketData, PacketDataResponse};
+use cw_ibc_core::ics04_channel::{PacketData, PacketDataResponse};
+use cw_ibc_core::ics24_host::commitment::compute_packet_commitment;
 use cw_ibc_core::types::ClientId;
 use ibc::core::ics03_connection::version::get_compatible_versions;
 use ibc::core::ics04_channel::msgs::timeout::MsgTimeout;
