+++ contracts/cosmwasm-vm/cw-ibc-core/tests/channel/test_channel_closeinit.rs
@@ -8,7 +8,6 @@ use std::{str::FromStr, time::Duration};
 use cosmwasm_std::{to_binary, Addr, Event, IbcOrder, Reply, SubMsgResponse, SubMsgResult};
 use cw_ibc_core::ics04_channel::open_init::create_channel_submesssage;
 use cw_ibc_core::ics04_channel::EXECUTE_ON_CHANNEL_CLOSE_INIT;
-use cw_ibc_core::traits::*;
 use cw_ibc_core::{
     context::CwIbcCoreContext,
     ics04_channel::MsgChannelCloseInit,
