+++ contracts/cosmwasm-vm/cw-ibc-core/tests/test_client.rs
@@ -20,18 +20,13 @@ use cw_ibc_core::{
     MsgCreateClient, MsgUpdateClient, MsgUpgradeClient,
 };
 use ibc::{
-    core::{
-        ics02_client::msgs::misbehaviour::MsgSubmitMisbehaviour,
-        ics23_commitment::commitment::CommitmentRoot,
-    },
+    core::ics02_client::msgs::misbehaviour::MsgSubmitMisbehaviour,
     mock::{
         client_state::MockClientState, consensus_state::MockConsensusState, header::MockHeader,
     },
     signer::Signer,
     Height,
 };
-
-use ibc_proto::ics23::CommitmentProof;
 use setup::*;
 
 #[test]
